requirements.txt
numpy
Pillow
PyOpenGL
PyOpenGL_accelerate
astropy
configparser
